requirements.txt
pandas
numpy
django
djangorestframework
